Reorganize JavaScript into modules
