+++ js/ranking.js
@@ -1,0 +1,52 @@
+import { ranquing, anySeleccionat, modalitatSeleccionada, appendResponsiveTable } from './init.js';
+import { mostraEvolucioJugador } from './graficos.js';
+
+export function mostraRanquing() {
+  const cont = document.getElementById('content');
+  cont.innerHTML = '';
+  const taula = document.createElement('table');
+  const cap = document.createElement('tr');
+  ['Posició', 'Jugador', 'Mitjana'].forEach(t => {
+    const th = document.createElement('th');
+    th.textContent = t;
+    cap.appendChild(th);
+  });
+  taula.appendChild(cap);
+
+  const dadesOrdenades = ranquing
+    .filter(
+      reg =>
+        parseInt(reg.Any, 10) === anySeleccionat &&
+        reg.Modalitat === modalitatSeleccionada
+    )
+    .sort((a, b) => parseFloat(b.Mitjana) - parseFloat(a.Mitjana));
+
+  dadesOrdenades.forEach((reg, idx) => {
+    const tr = document.createElement('tr');
+    ['Posició', 'Jugador', 'Mitjana'].forEach(clau => {
+      const td = document.createElement('td');
+      let valor;
+      if (clau === 'Posició') {
+        valor = idx + 1;
+      } else if (clau === 'Mitjana') {
+        valor = Number.parseFloat(reg[clau]).toFixed(3);
+      } else {
+        valor = clau === 'Jugador' ? reg.NomComplet : reg[clau];
+      }
+      td.textContent = valor;
+      if (clau === 'Jugador') {
+        td.classList.add('player-cell');
+        td.addEventListener('click', e => {
+          e.stopPropagation();
+          mostraEvolucioJugador(reg.Jugador, reg.NomComplet);
+        });
+      }
+      tr.appendChild(td);
+    });
+    tr.addEventListener('click', () => {
+      mostraEvolucioJugador(reg.Jugador, reg.NomComplet);
+    });
+    taula.appendChild(tr);
+  });
+  appendResponsiveTable(cont, taula);
+}
